Some fixes from PR review
