+++ Tests/APITesters/ObjCAPITester/ObjCAPITester/RCPresentedOfferingContextAPI.m
@@ -1,11 +1,11 @@
 //
-//  RCPresentedOfferingContext.m
+//  RCPresentedOfferingContextAPI.m
 //  ObjCAPITester
 //
 //  Created by Josh Holtz on 2/14/24.
 //
 
-#import "RCPresentedOfferingContext.h"
+#import "RCPresentedOfferingContextAPI.h"
 
 @import StoreKit;
 @import RevenueCat;
@@ -13,10 +13,12 @@
 @implementation RCPresentOfferingContextAPI
 
 + (void)checkAPI {
+    RCPresentedOfferingContext *poc1 = [[RCPresentedOfferingContext alloc] initWithOfferingIdentifier:@"" placementIdentifier:nil];
     RCPresentedOfferingContext *poc = [[RCPresentedOfferingContext alloc] initWithOfferingIdentifier:@"" placementIdentifier:@""];
     NSString *oid = poc.offeringIdentifier;
+    NSString *pid = poc.placementIdentifier;
 
-    NSLog(poc, oid);
+    NSLog(poc, oid, pid);
 }
 
 @end
